Rename "Matrix" to "ToneMatrix" for more consistency.
