+++ src/app/audio/audio.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AudioContextService } from './audio-context.service';
-import { MatrixLoopService } from './matrix-loop.service';
 
 @NgModule({
   imports: [
     CommonModule
   ],
   declarations: [],
-  providers: [AudioContextService, MatrixLoopService]
+  providers: [AudioContextService]
 })
 export class AudioModule { }
