+++ src/app/matrix/matrix.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ToneMatrixActions } from './actions';
+import { MatrixLoopService } from './matrix-loop.service';
 
 @NgModule({
   imports: [
     CommonModule
   ],
   declarations: [],
-  providers: [ToneMatrixActions]
+  providers: [ToneMatrixActions, MatrixLoopService]
 })
 export class MatrixModule { }
