+++ src/app/matrix/tone-matrix-rdx.ts
@@ -1,4 +1,4 @@
-export interface ToneMatrixRdx {
+export interface ToneMatrix {
     notes: Beat[]
 }
